% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script - run_snppit.R
\name{format.snpptFile}
\alias{format.snpptFile}
\title{SNPPITfile}
\usage{
\method{format}{snpptFile}(
  name,
  df_parents,
  df_offspring,
  parentSex = T,
  parentPopulation = T,
  parentGroup = T
)
}
\description{
SNPPITfile
}
\keyword{internal}
