% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script - run_snppit.R
\name{make_snppit_file}
\alias{make_snppit_file}
\title{Make snppit popgen file}
\usage{
make_snppit_file(
  df_offspring,
  df_parents,
  projectName = "undefined_project",
  useGroups = F
)
}
\description{
functions similarily to run_snppit, but only creates the popgen file and does not run snppit.exe
}
