% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script - run_snppit.R
\name{renameGenotypes}
\alias{renameGenotypes}
\title{Rename genotypes based on a lookup table}
\usage{
renameGenotypes(dataframe, LUT, not_genotypes = c())
}
\description{
In a dataframe, rename genotype columns
}
