% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script - run_snppit.R
\name{run_snppit}
\alias{run_snppit}
\title{Parentage analysis (using SNPPT.exe)}
\usage{
run_snppit(
  df_offspring,
  df_parents,
  projectName = "project1",
  overwrite = F,
  useGroups = F,
  full = F,
  run = T
)
}
\arguments{
\item{df_offspring}{Data frame contaning offspring ID's and SNP genotypes (see below)}

\item{df_parents}{Data frame containing parent ID's, sex, population, and SNP genotypes (see above)}

\item{projectName}{Optional. A name that will be used for files created during the proces}

\item{overwrite}{Optional. If the script should overwrite any older snppit analysis (or read the old one again instead of re-doing it)}

\item{useGroups}{Optional. If parent grups should be used, only parents within the same group will be considered as parents together.}

\item{full}{Optional. Set to T if you want the full snppit result file, otherwise you get a simplified one.}

\item{run}{Optional. Set to T if you don't want to run snppit but only create the popgen file}
}
\description{
Function that simplifies using SNPPT.exe to calculate parentage based on SNP data
Requires that snppt.exe is installed in the working directory of this file!
More info at: https://swfsc.noaa.gov/textblock.aspx?Division=FED&ParentMenuId=54&id=16021
}
\details{
Download snppit here: https://github.com/eriqande/snppit
Extract the zip-file, rename the "snppit-master" folder to "snppit", and place the folder in the same working directory as this script.
Then you're ready! Usually, to run snppit, you need to prepare a specially formatted genpop file with your offspring and parent SNP data,
and then run snppit via the command line. This r function saves you some time by doing the genpop formatting and command line stuff for you in R.
All you need to do is to supply the function with two datasets: parents and offspring, and the function will return the resulting parentage analysis ready to use in R.
The function takes two datasets as parameters, one for offspring and one for parents. \cr
Both datasets must be formated as folows:
\itemize{
\item Individuals as rows
\item SNP data on columns, one column pr SNP. Important: any column that is not named "ID", (or for parents, "sex" or "population"), will be interpreted as a SNP.
\item SNP genotype written numerically: 1=homozygous for allele a, 2=heterozygous, 3=homozygous for allele b
\item One column named "ID" that contains some unique ID for each individual,
\item Only for parents: One column named "sex" that's either "M" or "F"
\item Only for parents: One column named "population" that gives that individual's population
\item Only for parents: One column named "group" that gives the parents spawning group (foursome) must come after "sex"
}
}
\examples{
parentage_data <- run_snppt(offspring, parents, "Project_oct2019")
}
