% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script - run_snppit.R
\name{SNPPITfile}
\alias{SNPPITfile}
\title{SNPPITfile}
\usage{
SNPPITfile(
  name,
  parents,
  offspring,
  parentSex = T,
  parentPopulation = T,
  parentGroup = T
)
}
\description{
SNPPITfile
}
\keyword{internal}
