% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadesc-families.R
\docType{data}
\name{table_families}
\alias{table_families}
\title{family information}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 121 rows and 12 columns.
}
\usage{
data(table_families)
}
\description{
example data for run_snppit
family structures (pairings and other info)
}
\keyword{datasets}
