% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadesc-offspringgenotypes.R
\docType{data}
\name{table_offspring_genotypes}
\alias{table_offspring_genotypes}
\title{offspring SNP genotypes}
\format{
An object of class \code{data.frame} with 6199 rows and 176 columns.
}
\usage{
data(table_offspring_genotypes)
}
\description{
example data for run_snppit
SNP genotypes of offspring individuals
}
\keyword{datasets}
