% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadesc-parentgenotypes.R
\docType{data}
\name{table_parents_genotypes}
\alias{table_parents_genotypes}
\title{Parent SNP genotypes}
\format{
An object of class \code{data.frame} with 120 rows and 154 columns.
}
\usage{
data(table_parents_genotypes)
}
\description{
example data for run_snppit
SNP genotypes of parent individuals
}
\keyword{datasets}
